requirements.txt
pyswip
python-dotenv